+++ editor/editor-core/src/__tests__/visual-regression/media/selection.ts
@@ -137,7 +137,8 @@ describe('Snapshot Test: Media', () => {
         await pressKey(page, 'ArrowUp');
       });
 
-      it('should render right side gap cursor (via arrow left)', async () => {
+      // FIXME: This test was automatically skipped due to failure on 8/24/2021: https://product-fabric.atlassian.net/browse/ED-13662
+      it.skip('should render right side gap cursor (via arrow left)', async () => {
         await pressKey(page, 'ArrowLeft');
       });